backend/requirements.txt
django>=5.2.8
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.1
PyMySQL>=1.1.1
django-cors-headers>=4.4.0
tzdata>=2024.1
# Optional AI (install when needed)
sentence-transformers>=3.2.1 ; platform_system != 'Windows' or python_version >= '3.10'
scikit-learn>=1.5.2